requirements.txt
arrow==0.17.0
blessed==1.17.11
boto3==1.17.10
botocore==1.20.10
Django==3.0.5
django-extensions==2.2.3
django-storages==1.11.1
django-picklefield==3.0.1
django-q==1.3.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
Pillow==7.1.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.1
pytz==2019.2
six==1.12.0
s3transfer==0.3.4
sqlparse==0.3.0
wcwidth==0.2.5
cent==3.0.1
whitenoise==4.1.4
dj-database-url==0.5.0
